server/requirements.txt
openai-whisper==20231117
torch>=2.1.0
ffmpeg-python>=0.2.0
faster-whisper>=1.0.0






